#8 [refactor] : rename dto -> response
