+++ app/src/main/java/org/sopt/and/data/model/response/ResponseUserHobbyDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.and.data.model.dto
+package org.sopt.and.data.model.response
 
 
 import kotlinx.serialization.SerialName
